+++ src/opstools/cli.py
@@ -4,8 +4,8 @@ Top level command for opstools
 
 import logging
 import click
-import os
-import sys
+
+from terraform_cloud_deployer import __version__
 
 __author__ = "Afraz Ahmadzadeh"
 __copyright__ = "Afraz Ahmadzadeh"